requirements.txt
https://www.djangoproject.com/download/1.7.b4/tarball/
pillow==2.4.0
django-multiuploader==0.2.9
